+++ basic.h
@@ -6,8 +6,8 @@ Hadamard乘法 : VDOT(矩陣A, 矩陣B)
 #ifndef BASIC_H
 #define BASIC_H
 
-int** VADD(int A[3][3], int B[3][3]);
-int** VSUB(int A[3][3], int B[3][3]);
-int** VDOT(int A[3][3], int B[3][3]);
+double** VADD(double A[3][3], double B[3][3]);
+double** VSUB(double A[3][3], double B[3][3]);
+double** VDOT(double A[3][3], double B[3][3]);
 
 #endif
